# chore: replace `resolver-oxc` with new `resolver-typescript`