<!--

DO NOT INGORE THE TEMPLATE!
请不要忽视这个模板！

Thank you for contributing!
感谢你的贡献！

Before submitting the PR, please make sure you do the following:
在提交PR之前，请确保你做到以下几点：

- Check that there isn't already a PR that solves the problem the same way to avoid creating a duplicate.
- Provide a description in this PR that addresses **what** the PR is solving, or reference the issue that it solves (e.g. `fixes #123`).
- Ideally, include relevant tests that fail without this PR but pass with it.
- 检查是否已经有一个以同样方式解决该问题的 PR，以避免重复创建。
- 在这个 PR 中描述 **PR 所要解决的问题**，或者引用它所解决的问题（例如 `fixes #123`）。
- 理想情况下，提交没有这个 PR 的情况下失败但在有 PR 的情况下通过的相关测试。

-->

### Description 描述

<!--

Please insert your description here and provide especially info about the "what" this PR is solving
请在此插入你的描述，并提供有关该 PR 所解决的 **问题** 的信息。

-->

Hi, `eslint-plugin-import-x` and `eslint-import-resolver-typescript` owner here.

The brand new `eslint-import-resolver-typescript` is now powered by [unrs-resolver](https://github.com/unrs/unrs-resolver) which is fork of [oxc-resolver](https://github.com/oxc-project/oxc-resolver) and [rspack-resolver](https://github.com/web-infra-dev/rspack-resolver) with P'n'P support and more targets, and it has competitive performance compared with [`eslint-import-resolver-oxc`](https://github.com/9romise/eslint-import-resolver-oxc).

### Linked Issues 关联的 Issues

N/A

### Additional context 额外上下文

<!--

e.g. is there anything you'd like reviewers to focus on?
例如，有什么东西是你希望审查人关注的？

-->

Before with `resolver-oxc`:

```sh
 fabric on  main ❯ TIMING=1 p lint:eslint

> @modyqyw/fabric@11.8.1 lint:eslint /Users/JounQin/Workspaces/GitHub/fabric
> eslint . --fix --cache


/Users/JounQin/Workspaces/GitHub/fabric/src/cli.ts
    9:8   warning  Using exported name 'consola' as identifier for default export                      import/no-named-as-default
   71:7   warning  Unsafe assignment of an `any` value                                                 @typescript-eslint/no-unsafe-assignment
   87:35  warning  Unexpected any. Specify a different type                                            @typescript-eslint/no-explicit-any
  290:7   warning  Unsafe assignment of an `any` value                                                 @typescript-eslint/no-unsafe-assignment
  291:28  warning  Unsafe member access .scripts on an `any` value                                     @typescript-eslint/no-unsafe-member-access
  451:9   warning  Unsafe assignment of an `any` value                                                 @typescript-eslint/no-unsafe-assignment
  472:55  warning  'f' is already declared in the upper scope on line 471 column 21                    no-shadow
  479:49  warning  Unsafe member access [o.field] on an `any` value                                    @typescript-eslint/no-unsafe-member-access
  488:31  warning  Unexpected any. Specify a different type                                            @typescript-eslint/no-explicit-any
  496:32  warning  Unsafe member access .devDependencies on an `any` value                             @typescript-eslint/no-unsafe-member-access
  497:32  warning  Unsafe member access .dependencies on an `any` value                                @typescript-eslint/no-unsafe-member-access
  514:11  warning  Unsafe assignment of an `any` value                                                 @typescript-eslint/no-unsafe-assignment
  514:29  warning  Unsafe member access .scripts on an `any` value                                     @typescript-eslint/no-unsafe-member-access
  515:34  warning  Unsafe member access .scripts on an `any` value                                     @typescript-eslint/no-unsafe-member-access
  577:13  warning  Prefer top-level await over using a promise chain                                   unicorn/prefer-top-level-await
  577:19  warning  Avoid callbacks. Prefer Async/Await                                                 promise/prefer-await-to-callbacks
  579:19  warning  Unsafe argument of type `any` assigned to a parameter of type `string | undefined`  @typescript-eslint/no-unsafe-argument

✖ 17 problems (0 errors, 17 warnings)

Rule                                    | Time (ms) | Relative
:---------------------------------------|----------:|--------:
unicorn/no-unnecessary-polyfills        |   326.636 |    11.7%
@typescript-eslint/no-unsafe-assignment |   257.082 |     9.2%
@typescript-eslint/no-misused-promises  |   248.984 |     8.9%
tailwindcss/classnames-order            |   219.827 |     7.9%
n/no-extraneous-import                  |   188.172 |     6.8%
n/no-unpublished-import                 |   116.917 |     4.2%
unused-imports/no-unused-imports        |   103.141 |     3.7%
@typescript-eslint/no-floating-promises |    92.079 |     3.3%
import/no-self-import                   |    84.932 |     3.0%
import/no-named-as-default              |    76.287 |     2.7%
is 📦 v11.8.1 via  v22.14.0 took 16s
```

After with `resolver-typescript`

```sh
 fabric on  chore/import_resolver [+] ❯ TIMING=1 p lint:eslint

> @modyqyw/fabric@11.8.1 lint:eslint /Users/JounQin/Workspaces/GitHub/fabric
> eslint . --fix --cache


/Users/JounQin/Workspaces/GitHub/fabric/src/cli.ts
    9:8   warning  Using exported name 'consola' as identifier for default export                      import/no-named-as-default
   71:7   warning  Unsafe assignment of an `any` value                                                 @typescript-eslint/no-unsafe-assignment
   87:35  warning  Unexpected any. Specify a different type                                            @typescript-eslint/no-explicit-any
  290:7   warning  Unsafe assignment of an `any` value                                                 @typescript-eslint/no-unsafe-assignment
  291:28  warning  Unsafe member access .scripts on an `any` value                                     @typescript-eslint/no-unsafe-member-access
  451:9   warning  Unsafe assignment of an `any` value                                                 @typescript-eslint/no-unsafe-assignment
  472:55  warning  'f' is already declared in the upper scope on line 471 column 21                    no-shadow
  479:49  warning  Unsafe member access [o.field] on an `any` value                                    @typescript-eslint/no-unsafe-member-access
  488:31  warning  Unexpected any. Specify a different type                                            @typescript-eslint/no-explicit-any
  496:32  warning  Unsafe member access .devDependencies on an `any` value                             @typescript-eslint/no-unsafe-member-access
  497:32  warning  Unsafe member access .dependencies on an `any` value                                @typescript-eslint/no-unsafe-member-access
  514:11  warning  Unsafe assignment of an `any` value                                                 @typescript-eslint/no-unsafe-assignment
  514:29  warning  Unsafe member access .scripts on an `any` value                                     @typescript-eslint/no-unsafe-member-access
  515:34  warning  Unsafe member access .scripts on an `any` value                                     @typescript-eslint/no-unsafe-member-access
  577:13  warning  Prefer top-level await over using a promise chain                                   unicorn/prefer-top-level-await
  577:19  warning  Avoid callbacks. Prefer Async/Await                                                 promise/prefer-await-to-callbacks
  579:19  warning  Unsafe argument of type `any` assigned to a parameter of type `string | undefined`  @typescript-eslint/no-unsafe-argument

✖ 17 problems (0 errors, 17 warnings)

Rule                                    | Time (ms) | Relative
:---------------------------------------|----------:|--------:
unicorn/no-unnecessary-polyfills        |   396.741 |    13.7%
@typescript-eslint/no-misused-promises  |   319.480 |    11.1%
@typescript-eslint/no-unsafe-assignment |   235.795 |     8.2%
tailwindcss/classnames-order            |   232.258 |     8.0%
n/no-extraneous-import                  |   195.101 |     6.7%
import/export                           |   133.657 |     4.6%
n/no-unpublished-import                 |   110.246 |     3.8%
import/no-named-as-default              |   101.658 |     3.5%
@typescript-eslint/no-unsafe-return     |    88.166 |     3.1%
@typescript-eslint/no-floating-promises |    77.325 |     2.7%
is 📦 v11.8.1 via  v22.14.0 took 11s
```


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Chores**
  - Updated linting dependencies and configurations to ensure consistency.
- **Refactor**
  - Reorganized linting settings with clearer naming conventions for improved maintainability.
- **Documentation**
  - Enhanced and restructured the linting guide for better clarity, including localized materials.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->